Update litmus tests to be parsed on the latest MLIR

+++ tests/litmus/abstraction/dot_concat.tgt.mlir
@@ -3,8 +3,8 @@ func.func @f(%a: tensor<5xf32>, %b: tensor<5xf32>, %c: tensor<5xf32>, %d: tensor
   %i = tensor.empty (): tensor<f32>
   %outty = linalg.fill ins(%identity: f32) outs(%i: tensor<f32>) -> tensor<f32>
 
-  %ca = "tosa.concat"(%a, %c) {axis = 0: i64}: (tensor<5xf32>, tensor<5xf32>) -> tensor<10xf32>
-  %cb = "tosa.concat"(%b, %d) {axis = 0: i64}: (tensor<5xf32>, tensor<5xf32>) -> tensor<10xf32>
+  %ca = "tosa.concat"(%a, %c) {axis = 0: i32}: (tensor<5xf32>, tensor<5xf32>) -> tensor<10xf32>
+  %cb = "tosa.concat"(%b, %d) {axis = 0: i32}: (tensor<5xf32>, tensor<5xf32>) -> tensor<10xf32>
 
   %rt = linalg.dot ins(%ca, %cb : tensor<10xf32>, tensor<10xf32>)
       outs(%outty: tensor<f32>) -> tensor<f32>
